feat: [#173102366] Timeout activation screen (#1881)
* added timeoutActivationScreen

* [#173102366] fix merge conflicts
