+++ locales/en/index.yml
@@ -1265,8 +1265,11 @@ bonus:
         title: Sorry, it takes longer to verify your ISEE.
         description: "We will send you a message when INPS has given us the answer related to your ISEE.\n\nWhen you receive it, if it is positive, remember to complete your application within 24 hours, otherwise it will expire and you will have to start a new one."
       activate:
-        loading: "We have forwarded your request to Agenzia delle Entrate\nPlease wait."
-        goToDetails: View bonus
+        loading: Bonus activation
+        goToDetails: Show details
+        timeout:
+          title: Sorry, it takes longer to process your request.
+          description: We will send you a message in the app when your bonus is ready.
         completed:
           title: Congratulation!
           description: Bonus activated
