+++ locales/it/index.yml
@@ -1298,8 +1298,11 @@ bonus:
         title: Siamo spiacenti, serve più tempo per verificare il tuo ISEE.
         description: "Ti invieremo un messaggio in app quando INPS ci avrà fornito la risposta relativa al tuo ISEE.\n\nQuando lo riceverai, se sarà positivo, ricordati di completare la tua domanda entro 24 ore, altrimenti scadrà e dovrai iniziarne una nuova."
       activate:
-        loading: "Abbiamo inoltrato la tua richiesta a Agenzia delle Entrate\nTi preghiamo di attendere."
-        goToDetails: Visualizza bonus
+        loading: Attivo il bonus
+        goToDetails: Visualizza i dettagli
+        timeout:
+          title: Siamo spiacenti, serve più tempo per processare la tua richiesta.
+          description: Ti invieremo un messaggio in app quando il tuo bonus sarà pronto.
         completed:
           title: Congratulazioni!
           description: Il tuo bonus è attivo
