+++ ts/features/bonusVacanze/navigation/navigator.ts
@@ -6,10 +6,11 @@ import BonusInformationScreen from "../screens/BonusInformationScreen";
 import ActivateBonusCompletedScreen from "../screens/eligibility/ActivateBonus/ActivateBonusCompletedScreen";
 import LoadActivateBonusScreen from "../screens/eligibility/ActivateBonus/LoadActivateBonusScreen";
 import ActivateBonusRequestScreen from "../screens/eligibility/ActivateBonus/request/ActivateBonusRequestScreen";
+import { TimeoutActivationInfoScreen } from "../screens/eligibility/ActivateBonus/TimeoutActivationInfoScreen";
 import IseeNotAvailableScreen from "../screens/eligibility/isee/IseeNotAvailableScreen";
 import IseeNotEligibleScreen from "../screens/eligibility/isee/IseeNotEligibleScreen";
 import LoadBonusEligibilityScreen from "../screens/eligibility/LoadBonusEligibilityScreen";
-import TimeoutEligibilityCheckInfoScreen from "../screens/eligibility/TimeoutEligibilityCheckInfoScreen";
+import { TimeoutEligibilityCheckInfoScreen } from "../screens/eligibility/TimeoutEligibilityCheckInfoScreen";
 import BONUSVACANZE_ROUTES from "./routes";
 
 const BonusVacanzeNavigator = createStackNavigator(
@@ -43,6 +44,9 @@ const BonusVacanzeNavigator = createStackNavigator(
     },
     [BONUSVACANZE_ROUTES.ELIGIBILITY.ACTIVATION_LOADING]: {
       screen: LoadActivateBonusScreen
+    },
+    [BONUSVACANZE_ROUTES.ELIGIBILITY.ACTIVATION_TIMEOUT]: {
+      screen: TimeoutActivationInfoScreen
     },
     [BONUSVACANZE_ROUTES.ELIGIBILITY.ACTIVATION_COMPLETED]: {
       screen: ActivateBonusCompletedScreen
