+++ ts/features/bonusVacanze/navigation/routes.ts
@@ -14,6 +14,7 @@ const BONUSVACANZE_ROUTES = {
     TIMEOUT: "BONUS_ELIGIBILITY_TIMEOUT",
     ACTIVATION_REQUEST: "BONUS_ELIGIBILITY_ACTIVATION_REQUEST",
     ACTIVATION_LOADING: "BONUS_ELIGIBILITY_ACTIVATION_LOADING",
+    ACTIVATION_TIMEOUT: "BONUS_ELIGIBILITY_ACTIVATION_TIMEOUT",
     ACTIVATION_COMPLETED: "BONUS_ELIGIBILITY_ACTIVATION_COMPLETED",
     ACTIVATE_BONUS: "BONUS_ELIGIBILITY_ACTIVATE"
   }
